flask_backend/requirements.txt
Flask==2.3.3
Flask-SocketIO==5.3.6
Flask-Cors==4.0.0
python-socketio==5.10.0
Werkzeug==2.3.7
python-dotenv==1.0.0
Pillow==10.0.0
opencv-python==4.8.0.74
pytesseract==0.3.10
numpy==1.24.3
matplotlib==3.7.2
dlib>=19.7
face_recognition==1.3.0
# eventlet==0.33.3  # Comment out eventlet
gevent==23.9.1  # Add gevent instead
gevent-websocket==0.10.1  # Add gevent-websocket for WebSocket support
# Install face_recognition_models with:
# pip install git+https://github.com/ageitgey/face_recognition_models




